# Recursive instruction planner