--- a/common/types/map.go
+++ b/common/types/map.go
@@ -28,9 +28,22 @@ type baseMap struct {
 	refValue reflect.Value
 }
 
+type stringMap struct {
+	*baseMap
+	mapStrStr map[string]string
+}
+
 // NewDynamicMap returns a traits.Mapper value with dynamic key, value pairs.
 func NewDynamicMap(value interface{}) traits.Mapper {
 	return &baseMap{value, reflect.ValueOf(value)}
+}
+
+// NewStringStringMap returns a specialized traits.Mapper with string keys and values.
+func NewStringStringMap(value map[string]string) traits.Mapper {
+	return &stringMap{
+		baseMap:   &baseMap{value: value},
+		mapStrStr: value,
+	}
 }
 
 var (
@@ -104,6 +117,13 @@ func (m *baseMap) ConvertToNative(refType reflect.Type) (interface{}, error) {
 	return nativeMap.Interface(), nil
 }
 
+func (m *stringMap) ConvertToNative(refType reflect.Type) (interface{}, error) {
+	if !m.baseMap.refValue.IsValid() {
+		m.baseMap.refValue = reflect.ValueOf(m.value)
+	}
+	return m.baseMap.ConvertToNative(refType)
+}
+
 func (m *baseMap) ConvertToType(typeVal ref.Type) ref.Value {
 	switch typeVal {
 	case MapType:
@@ -136,6 +156,13 @@ func (m *baseMap) Equal(other ref.Value) ref.Value {
 	return True
 }
 
+func (m *stringMap) Equal(other ref.Value) ref.Value {
+	if !m.baseMap.refValue.IsValid() {
+		m.baseMap.refValue = reflect.ValueOf(m.value)
+	}
+	return m.baseMap.Equal(other)
+}
+
 func (m *baseMap) Get(key ref.Value) ref.Value {
 	thisKeyType := m.refValue.Type().Key()
 	nativeKey, err := key.ConvertToNative(thisKeyType)
@@ -153,6 +180,18 @@ func (m *baseMap) Get(key ref.Value) ref.Value {
 	return NativeToValue(value.Interface())
 }
 
+func (m *stringMap) Get(key ref.Value) ref.Value {
+	strKey, ok := key.(String)
+	if !ok {
+		return ValOrErr(key, "no such key: %v", key)
+	}
+	val, found := m.mapStrStr[string(strKey)]
+	if !found {
+		return NewErr("no such key: %s", key)
+	}
+	return String(val)
+}
+
 func (m *baseMap) Iterator() traits.Iterator {
 	mapKeys := m.refValue.MapKeys()
 	return &mapIterator{
@@ -161,6 +200,13 @@ func (m *baseMap) Iterator() traits.Iterator {
 		mapKeys:      mapKeys,
 		cursor:       0,
 		len:          int(m.Size().(Int))}
+}
+
+func (m *stringMap) Iterator() traits.Iterator {
+	if !m.baseMap.refValue.IsValid() {
+		m.baseMap.refValue = reflect.ValueOf(m.value)
+	}
+	return m.baseMap.Iterator()
 }
 
 func (m *baseMap) Size() ref.Value {