The recursive instruction planner walks the `Expr` AST to generate an `Interpretable` call
tree whose branches represent linear evaluation paths. 

This approach offers the following advantages over prior approaches:

* Thread-safe evaluation that leverages `Go` call frames.
* Reduced evaluation overhead by removed a custom alloc to track eval state.
* Specialization of the `&&`, `||`, `? :`, and (in)equality operators.

The improvements in execution are obvious from the benchmarks:

```
benchmark                                       old ns/op     new ns/op     delta
BenchmarkInterpreter/ExprBench/ok_1st-12        615           128           -79.19%
BenchmarkInterpreter/ExprBench/ok_2nd-12        1787          531           -70.29%
BenchmarkInterpreter/ExprBench/not_found-12     1778          535           -69.91%
BenchmarkInterpreter/ExprBench/false_2nd-12     309           11.6          -96.25%

benchmark                                       old allocs     new allocs     delta
BenchmarkInterpreter/ExprBench/ok_1st-12        1              1              +0.00%
BenchmarkInterpreter/ExprBench/ok_2nd-12        5              4              -20.00%
BenchmarkInterpreter/ExprBench/not_found-12     5              4              -20.00%
BenchmarkInterpreter/ExprBench/false_2nd-12     0              0              +0.00%

benchmark                                       old bytes     new bytes     delta
BenchmarkInterpreter/ExprBench/ok_1st-12        8             8             +0.00%
BenchmarkInterpreter/ExprBench/ok_2nd-12        104           88            -15.38%
BenchmarkInterpreter/ExprBench/not_found-12     104           88            -15.38%
BenchmarkInterpreter/ExprBench/false_2nd-12     0             0             +0.00%

```

This PR also includes the following functionality changes:

* Functional `InterpretableDecorator` values may be provided at `Interpretable` creation time
  to layer or replace default `Interpretable` behavior. See `interpreter.ExhaustiveEval`, and
  `interpreter.FoldConstants`.
* `Interpretable` creation now considers the `exprpb.CheckedExpr` metadata to resolve namespaced
   identifiers within `Select` expressions to enum and identifier references.
* Equality checks now consider types inequality to be an error rather than `false`. #155 
* Specialized `map[string]string` handling for field accesses.

This PR replaces #152.