@TristonianJones 2019-01-28
Closes #159